pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-acp-reference-sep"
version = "0.1.0"
description = "Reference implementation for MCP Action-Level Authorization (ACP + TAM)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.2.0",
    "python-jose>=3.3.0",
    "httpx>=0.27.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-mcp-adapters>=0.1.0",
    "starlette>=0.39.0",
    "uvicorn>=0.30.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = [
    "mcp_server",
    "middleware",
    "examples",
]

